Changed UsersService to UserService, its path from users to user, getAll to list, getOne to retrieve; Updated imports to new path and class; Added docstrings to User related services, model, enums, resolver;
