+++ web/src/app/core/user/settings/settings.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 
-import { User } from 'src/app/shared/services/users/users.model';
-import { UsersService } from 'src/app/shared/services/users/users.service';
+import { User } from 'src/app/shared/services/user/user.model';
+import { UserService } from 'src/app/shared/services/user/user.service';
 
 @Component({
   selector: 'app-settings',
@@ -15,7 +15,7 @@ export class SettingsComponent implements OnInit, OnDestroy {
   accountInfo: User | undefined
 
   constructor(
-    private userSvc: UsersService
+    private userSvc: UserService
   ) { }
 
   ngOnInit(): void {
