+++ web/src/app/core/user/ticket-detail/ticket-detail.component.ts
@@ -13,10 +13,10 @@ import {
   TicketPriority,
   TicketStatus } from 'src/app/shared/services/tickets/tickets.model';
 import { UnitNo } from 'src/app/shared/services/units/units.model';
-import { UserEmail } from 'src/app/shared/services/users/users.model';
+import { UserEmail } from 'src/app/shared/services/user/user.model';
 import { TicketsService } from 'src/app/shared/services/tickets/tickets.service';
 import { UnitsService } from 'src/app/shared/services/units/units.service';
-import { UsersService } from 'src/app/shared/services/users/users.service';
+import { UserService } from 'src/app/shared/services/user/user.service';
 import { AddTicketComponent } from 'src/app/components/tickets/add-ticket/add-ticket.component';
 
 @Component({
@@ -88,7 +88,7 @@ export class TicketDetailComponent implements OnInit, OnDestroy {
     private helper: HelpersService,
     private ticketSvc: TicketsService,
     private unitSvc: UnitsService,
-    private userSvc: UsersService
+    private userSvc: UserService
   ) { }
 
   ngOnInit(): void {
