+++ web/src/app/shared/services/activities/activities.model.ts
@@ -1,5 +1,5 @@
 import { UnitNo } from '../units/units.model';
-import { UserEmail } from '../users/users.model';
+import { UserEmail } from '../user/user.model';
 
 export class UnitActivity {
     constructor(
