+++ web/src/app/shared/services/user/user.service.ts
@@ -4,14 +4,17 @@ import { Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
 import { environment } from 'src/environments/environment';
 
-import { User, EmailVerification, UserEmail } from './users.model';
+import { User, EmailVerification, UserEmail } from './user.model';
 
 const BASE_URL = `${ environment.baseUrl }v1/users/`
 
+/**
+ * A service for User related methods
+ */
 @Injectable({
   providedIn: 'root'
 })
-export class UsersService {
+export class UserService {
 
   // Data
   currentUser: User | undefined
@@ -24,7 +27,12 @@ export class UsersService {
     private http: HttpClient
   ) { }
 
-  getAll(): Observable<User[]> {
+  /**
+   * Returns a list of users
+   *
+   * @returns List of users
+   */
+  list(): Observable<User[]> {
     const urlTemp = `${ BASE_URL }`
     return this.http.get<User[]>(urlTemp).pipe(
       tap((res: User[]) => {
@@ -34,7 +42,14 @@ export class UsersService {
     )
   }
 
-  getOne(id: number): Observable<User> {
+  /**
+   * Returns a user
+   * 
+   * @param id - user ID
+   *
+   * @returns A user
+   */
+  retrieve(id: number): Observable<User> {
     const urlTemp = `${ BASE_URL }${ id }/`
     return this.http.get<User>(urlTemp).pipe(
       tap((res: User) => {
@@ -44,6 +59,11 @@ export class UsersService {
     )
   }
 
+  /**
+   * Returns the current user information
+   *
+   * @returns Current user
+   */
   getAccountInfo(): Observable<User> {
     const urlTemp = `${ BASE_URL }get-account-info/`
     return this.http.get<User>(urlTemp).pipe(
@@ -54,6 +74,15 @@ export class UsersService {
     )
   }
 
+  /**
+   * Partial update a user
+   *
+   * @param id - user ID
+   * @param body - payload
+   * @param body.fullName - user's full name
+   * 
+   * @returns Updated user
+   */
   patchAccount(body: any, id: number): Observable<User> {
     const urlTemp = `${ BASE_URL }${id}/`
     return this.http.patch<User>(urlTemp, body).pipe(
@@ -64,6 +93,11 @@ export class UsersService {
     )
   }
 
+  /**
+   * Returns a list of users with verifications status
+   *
+   * @returns List of users
+   */
   getAllVerification(): Observable<EmailVerification[]> {
     const urlTemp = `${ BASE_URL }get-users-verification/`
     return this.http.get<EmailVerification[]>(urlTemp).pipe(
@@ -74,6 +108,15 @@ export class UsersService {
     )
   }
 
+  /**
+   * Filter users through:
+   * - userType
+   * - isActive
+   * 
+   * @param filterStr - string filter
+   *
+   * @returns List of filtered users
+   */
   filterSimplified(filterStr?: string): Observable<UserEmail[]> {
     const qStr = filterStr ? `?${filterStr}` : ''
     const urlTemp = `${ BASE_URL }get-simplified/${ qStr }`
